+++ openPER/ViewModels/ClicheDrawingViewModel.cs
@@ -1,12 +1,11 @@
 ﻿using System.Collections.Generic;
-using System.Security.AccessControl;
 
 namespace openPER.ViewModels
 {
     public class ClicheDrawingViewModel
     {
         public List<PartViewModel> Parts { get; set; }
-        public double ParentPartNumber { get; set; } 
+        public decimal ParentPartNumber { get; set; } 
         public int CurrentDrawingNumber { get; set; }
     }
 }
